vrouter: use uint64_t in vr_get_time and vr_get_mono_time
Changes required to support Contrail on Windows.
- use uint64_t in vr_get_time, instead of unsigned long
  (which is 32 bit on Windows)
- use uint64_t in vr_get_mono_time, instead of unsigned long
  (which is 32 bit on Windows)
- add missing __attribute__packed__open__ on vr_pbb_itag
- remove forward declaration of rtable_families in dp-core/vr_route.c
  (MSVC does not support forward declaring an array without a size)
- miscellaneous compilation fixes

Partial-Bug: #1734699

+++ freebsd/vrouter_mod.c
@@ -369,7 +369,7 @@ fh_set_defer_data(void *data)
 }
 
 static void
-fh_get_time(unsigned long *sec, unsigned long *usec)
+fh_get_time(uint64_t *sec, uint64_t *usec)
 {
 	struct timespec tsp;
 
@@ -382,7 +382,7 @@ fh_get_time(unsigned long *sec, unsigned long *usec)
 }
 
 static void
-fh_get_mono_time(unsigned int *sec, unsigned int *nsec)
+fh_get_mono_time(uint64_t *sec, uint64_t *nsec)
 {
 	struct timespec tsp;
